Extracted CanvasExtensions into its own file. Moved CanvasExtensions into a new namespace.
