+++ AnnoMapEditor/UI/Controls/MapObject.xaml.cs
@@ -7,6 +7,7 @@ using System.Windows.Media;
 using System.Windows.Input;
 using System.Windows.Shapes;
 using System.Windows.Media.Imaging;
+using AnnoMapEditor.UI.Utilities;
 
 namespace AnnoMapEditor.UI.Controls
 {
